Svc REST: Remove old rest_test
All the tests have been ported to storage_test.go
